\name{moduleColor.revisionDate}
\alias{moduleColor.revisionDate}
\title{Get the last revision date of the package.}
\description{
  Returns the last revision date of the package.
}
\usage{
moduleColor.revisionDate()
}
\value{
  A character string.
}
\author{
  Peter Langfelder, \email{Peter.Langfelder@gmail.com}
}
\keyword{misc}
