\name{moduleColor.version}
\alias{moduleColor.version}
\title{Get the version number of the package.}
\description{
  Returns the version number of the package.
}
\usage{
moduleColor.version()
}
\value{
  A character string.
}
\author{
  Peter Langfelder, \email{Peter.Langfelder@gmail.com}
}
\keyword{misc}
